Rename `visualisation::Factory` to `visualisation::Class`.
