+++ src/rust/lib/graph-editor/src/component/visualization.rs
@@ -21,13 +21,11 @@
 pub mod class;
 pub mod container;
 pub mod data;
-pub mod factory;
 pub mod registry;
 pub mod renderer;
 
 pub use class::*;
 pub use container::*;
 pub use data::*;
-pub use factory::*;
 pub use registry::*;
 pub use renderer::*;
