+++ src/rust/lib/graph-editor/src/component/visualization/class.rs
@@ -5,7 +5,10 @@ use crate::prelude::*;
 use crate::frp;
 use crate::visualization::*;
 
+use ensogl::display::Scene;
 use ensogl::display;
+use std::error::Error;
+
 
 
 
@@ -171,5 +174,148 @@ impl Visualization {
     /// Set the viewport size of the visualization.
     pub fn set_size(&self, size:Vector2<f32>) {
         self.state.renderer.set_size(size)
+    }
+}
+
+
+
+
+// =============================
+// === Visualisation Class ===
+// =============================
+
+/// Result of the attempt to instantiate a `Visualisation` from a `Class`.
+pub type InstantiationResult = Result<Visualization,Box<dyn Error>>;
+
+
+/// Specifies a trait that allows the instantiation of `Visualisations`.
+///
+/// The `Class` provides a way to implement structs that allow the instantiation of specific
+/// visualisations, while already providing general information that doesn't require an
+/// instantiated visualization, for example, the name or input type of the visualisation.
+///
+/// There are two example implementations: The `JsSourceClass`, which is based on a JS snippet to
+/// instantiate `JsRenderer`, and the fairly generic `NativeConstructorClass`, that only requires
+/// a function that can create a InstantiationResult. The later can be used as a thin wrapper around
+/// the constructor methods of native visualizations.
+///
+/// Example
+/// --------
+/// ```
+/// use graph_editor::component::visualization;
+/// use graph_editor::component::visualization::Visualization;
+/// use graph_editor::component::visualization::renderer::example::native::BubbleChart;
+/// use ensogl::display::Scene;
+/// use std::rc::Rc;
+///
+/// // Create a `visualisation::Class` from a JS source code snippet.
+/// let js_source_class = visualization::JsSourceClass::from_js_source_raw(r#"
+///
+///    class BubbleVisualisation {
+///         onDataReceived(root, data) {}
+///         setSize(root, size) {}
+///     }
+///
+///     return new BubbleVisualisation();
+/// "#.into());
+///
+/// // Create a `visualisation::Class` that instantiates a `BubbleChart`.
+/// let native_bubble_vis_class = visualization::NativeConstructorClass::new(
+///     visualization::Metadata {
+///         name        : "Bubble Visualisation (native)".to_string(),
+///         input_types : vec!["[[float;3]]".to_string().into()],
+///     },
+///     Rc::new(|scene:&Scene| Ok(Visualization::new(BubbleChart::new(scene))))
+/// );
+/// ```
+pub trait Class: Debug {
+    /// Indicate which `DataType`s can be rendered by this visualization.
+    fn metadata(&self) -> &Metadata;
+    /// Create new visualisation, that is initialised for the given scene. This can fail if the
+    /// `visualisation::Class` contains invalid data, for example, JS code that fails to execute, of if the
+    /// scene is in an invalid state.
+    // TODO consider not allowing failing here and require the checking on instantiation of the `Class`.
+    fn instantiate(&self, scene:&Scene) -> InstantiationResult;
+}
+
+
+
+// =======================
+// === JS Source Class ===
+// =======================
+
+#[derive(CloneRef,Clone,Debug)]
+#[allow(missing_docs)]
+pub struct JsSourceClass {
+    info   : Rc<Metadata>,
+    source : Rc<CowString>,
+}
+
+impl JsSourceClass {
+    /// Create a visualisation source from piece of JS source code and some metadata.
+    pub fn from_js_source(info:Metadata, source:CowString) -> Self {
+        let info   = Rc::new(info);
+        let source = Rc::new(source);
+        JsSourceClass {info,source}
+    }
+
+    /// Create a visualisation source from piece of JS source code. Metadata needs to be inferred.
+    pub fn from_js_source_raw(source:CowString) -> Self {
+        // TODO specify a way to provide this information fom raw source files.
+        let info  = Rc::new(Metadata {
+            name: "Unknown".to_string(),
+            input_types: vec![]
+        });
+        let source = Rc::new(source);
+        JsSourceClass {info,source}
+    }
+}
+
+impl Class for JsSourceClass {
+    fn metadata(&self) -> &Metadata {
+        &self.info
+    }
+
+    fn instantiate(&self, scene:&Scene) -> InstantiationResult {
+        let renderer = JsRenderer::from_constructor(&self.source)?;
+        renderer.set_dom_layer(&scene.dom.layers.front);
+        Ok(Visualization::new(renderer))
+    }
+}
+
+
+
+// ================================
+// === Native Constructor Class ===
+// ================================
+
+/// Type alias for a function that can create a `Visualisation`.
+pub type VisualisationConstructor = dyn Fn(&Scene) -> InstantiationResult;
+
+#[derive(CloneRef,Clone,Derivative)]
+#[derivative(Debug)]
+#[allow(missing_docs)]
+pub struct NativeConstructorClass {
+    info        : Rc<Metadata>,
+    #[derivative(Debug="ignore")]
+    constructor : Rc<VisualisationConstructor>,
+}
+
+impl NativeConstructorClass {
+    /// Create a visualisation source from a closure that returns a `Visualisation`.
+    pub fn new(info:Metadata, constructor:Rc<VisualisationConstructor>) -> Self {
+        let info = Rc::new(info);
+        NativeConstructorClass { info,constructor }
+    }
+}
+
+
+impl Class for NativeConstructorClass {
+    fn metadata(&self) -> &Metadata {
+        &self.info
+    }
+
+    fn instantiate(&self, scene:&Scene) -> InstantiationResult {
+        self.constructor.call((scene,))
     }
 }
