+++ src/rust/lib/graph-editor/src/lib.rs
@@ -357,7 +357,7 @@ pub struct FrpInputs {
     pub translate_selected_nodes       : frp::Source<Position>,
     pub cycle_visualization            : frp::Source<NodeId>,
     pub set_visualization              : frp::Source<(NodeId,Option<Visualization>)>,
-    pub register_visualisation_source  : frp::Source<Option<Rc<dyn visualization::Factory>>>,
+    pub register_visualisation_source  : frp::Source<Option<Rc<dyn visualization::Class>>>,
 }
 
 impl FrpInputs {
@@ -1478,7 +1478,7 @@ fn new_graph_editor(world:&World) -> GraphEditor {
 
         def _register_visualization = inputs.register_visualisation_source.map(f!((visualization_registry)(source) {
             if let Some(source) = source {
-                visualization_registry.register_factory_rc(source.clone_ref());
+                visualization_registry.register_class_rc(source.clone_ref());
             }
         }));
